bug fixes, refactor names

+++ src/graphql_sqlalchemy/resolvers.py
@@ -194,7 +194,7 @@ def make_delete_resolver(model: DeclarativeMeta) -> Callable:
 
         rows = query.all()
         affected = query.delete()
-        session.commit()
+        session_commit(session)
 
         return {"affected_rows": affected, "returning": rows}
 
@@ -202,14 +202,16 @@ def make_delete_resolver(model: DeclarativeMeta) -> Callable:
 
 
 def make_delete_by_pk_resolver(model: DeclarativeMeta) -> Callable:
-    def resolver(_root: None, info: Any, **kwargs: Dict[str, Any]) -> List[DeclarativeMeta]:
+    def resolver(_root: None, info: Any, **kwargs: Dict[str, Any]) -> Optional[List[DeclarativeMeta]]:
         session = info.context["session"]
 
         row = session.query(model).get(kwargs)
-        session.delete(row)
-        session.commit()
+        if row:
+            session.delete(row)
+            session_commit(session)
+            return row
 
-        return row
+        return None
 
     return resolver
 
@@ -239,13 +241,14 @@ def make_update_resolver(model: DeclarativeMeta) -> Callable:
         _root: None,
         info: Any,
         where: Dict[str, Any],
-        _set: Optional[Dict[str, Any]],
-        _inc: Optional[Dict[str, Any]],
+        _set: Optional[Dict[str, Any]] = None,
+        _inc: Optional[Dict[str, Any]] = None,
     ) -> Dict[str, Union[int, List[DeclarativeMeta]]]:
         session = info.context["session"]
         query = session.query(model)
         query = filter_query(model, query, where)
         affected = update_query(query, model, _set, _inc)
+        session_commit(session)
 
         return {
             "affected_rows": affected,
@@ -259,14 +262,15 @@ def make_update_by_pk_resolver(model: DeclarativeMeta) -> Callable:
     def resolver(
         _root: None,
         info: Any,
-        pk_columns: Dict[str, Any],
-        _set: Optional[Dict[str, Any]],
-        _inc: Optional[Dict[str, Any]],
+        _set: Optional[Dict[str, Any]] = None,
+        _inc: Optional[Dict[str, Any]] = None,
+        **pk_columns: Dict[str, Any],
     ) -> Optional[DeclarativeMeta]:
         session = info.context["session"]
         query = session.query(model).filter_by(**pk_columns)
 
         if update_query(query, model, _set, _inc):
+            session_commit(session)
             return query.one()
 
         return None
